requirements.txt
streamlit
pandas
openai
sqlalchemy
streamlit-audiorecorder
python-dotenv
